requirements.txt
seleniumbase
pillow
pytesseract
environs
rich
opencv-python
